requirements.txt
python-telegram-bot[asyncio]>=20.0
python-dotenv
supabase
httpx
asyncpg
